update FederationSdlPrinter to work with applied directives
`FederationSdlPrinter` is a copy of a `graphql.schema.idl.SchemaPrinter` from `graphql-java` v17 with some custom filtering logic that was required for Federation v1.

`graphql-java` v18 introduced concept of applied directive to make a distinction from the directive definition (previously same `GraphQLDirective` type was used for both). This PR updates our printer to the latest version that available in the `graphql-java` repository (which contains the printer fix for schema directives). Using custom predicates for filtering directives and schema elements we can replicate our custom functionality without the need of a custom printer. Once new version of `graphql-java` is released (v19) we should be able to drop our copy and just call the printer included in the library.

Other changes:
* Drop tests that verify SDL printer functionality. Since printer is a copy of `graphql-java` we don't need to verify functionality provided and tested by the underlying lib.